Jubbles now lock on and head toward other jubbles within a certain constant detection radius (currently 60px).
Jubble colours are initialised randomly.

3 jubbles are now the default starting configuration.

When two jubbles lock on to each other and eventually converge on each others'
positions, a bug occurs where the single double-jubble entity now moves as one,
and goes twice as fast as normal. This won't be a problem in the final product
because jubbles will enter fight-or-flight mode, and neither of these involve
sharing a body with the other jubble.