Jubbles now have a line-of-sight (default 180deg) defined by their current angle (visible as the direction their nose is pointing). A jubble is only capable of detecting points that are within a certain distance from it, AND that are within a 180deg angle from its nose.
An interesting side effect of the above is that now, when two jubbles head
toward each other, when they collide this counts as a 'goal reached' state and
they will start to random-walk from there. Previously, they'd permanently tail
each other.

An area which could use improvement is that often, a jubble will see another and
will begin turning around to chase it but will not be able to fully turn in time
and the other jubble will become undetectable again. Consider speeding up
turning when there's a jubble goal?

There are also some comp geom functions that are commented out at the bottom of
the code because I thought I'd need them but I didn't. They may come in useful
later one.